backend/requirements.txt
# WuwaAI Backend Requirements

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 游戏自动化框架
# ok-script @ git+https://github.com/ok-oldking/ok-script.git

# LLM & Vision
transformers>=4.36.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
accelerate>=0.25.0

# OCR
opencv-python>=4.8.0
numpy>=1.24.0
easyocr>=1.7.0

# 网络 & 工具
httpx>=0.26.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# 日志
colorlog>=6.8.0

# WebSocket
websockets>=12.0
